code factor
